battlefield/gsm_provider/requirements.txt
appnope==0.1.2
asgiref==3.3.1
backcall==0.2.0
decorator==4.4.2
Django==3.1.7
django-debug-toolbar==3.2
django-extensions==3.1.1
django-rest-framework==0.1.0
djangorestframework==3.12.2
Faker==6.5.0
freezegun==1.1.0
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
model-bakery==1.2.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
psycopg2-binary==2.8.6
ptyprocess==0.7.0
Pygments==2.8.0
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
traitlets==5.0.5
wcwidth==0.2.5
